feat: add tags functionality and restructure files for i18n support
- Added tags table and book2tag mapping table to support book tagging
- Updated  function to include tags in API response
- Adjusted file structure to align with locale-based routing for i18n compatibility
