+++ app/[locale]/components/Footer.tsx
