+++ app/[locale]/components/Header.tsx
