+++ app/[locale]/components/Logo.tsx
