+++ app/[locale]/layout.tsx
@@ -2,10 +2,10 @@ import {NextIntlClientProvider} from 'next-intl';
 import {getMessages} from 'next-intl/server';
 import {notFound} from 'next/navigation';
 import {Locale, routing} from '@/i18n/routing';
-
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import LayoutWrapper from './LayoutWrapper';
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -30,6 +30,8 @@ export default async function LocaleLayout({
   params: {locale: string};
 }) {
   const {locale} = await params;
+
+  
   // Ensure that the incoming `locale` is valid
   if (!routing.locales.includes(locale as Locale)) {
     notFound();
@@ -38,13 +40,16 @@ export default async function LocaleLayout({
   // Providing all messages to the client
   // side is the easiest way to get started
   const messages = await getMessages();
+
   return (
     <html lang={locale}>
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
         <NextIntlClientProvider messages={messages}>
-          {children}
+          <LayoutWrapper>
+            {children}
+          </LayoutWrapper>
         </NextIntlClientProvider>
       </body>
     </html>
