+++ app/[locale]/page.tsx
@@ -1,10 +1,10 @@
 import {useTranslations} from 'next-intl';
- 
-export default function HomePage() {
+
+export default function Home() {
   const t = useTranslations('HomePage');
   return (
-    <div>
-      <h1>{t('title')}</h1>
+    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
+      <h1 className="text-4xl font-bold text-center">Welcome to DreamazeBook</h1>
     </div>
   );
 }
