+++ app/[locale]/personalize/page.tsx
