--- app/layout.tsx
@@ -1,37 +1,0 @@
-'use client'
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
-import { useSelectedLayoutSegments } from 'next/navigation';
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  const segments = useSelectedLayoutSegments();
-  const isPersonalizePage = segments.includes("personalize");
-
-  return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {!isPersonalizePage && <Header />}
-        {children}
-        {!isPersonalizePage && <Footer />}
-      </body>
-    </html>
-  );
-}
