+++ next.config.ts
@@ -1,10 +1,10 @@
-import type { NextConfig } from "next";
+import { NextConfig } from 'next'
 import createNextIntlPlugin from 'next-intl/plugin';
 
 const withNextIntl = createNextIntlPlugin();
 
 /** @type {import('next').NextConfig} */
-const nextConfig = {
+const nextConfig: NextConfig = {
   /* config options here */
   reactStrictMode: true,
   images: {
